address issues per ohad and previous commit
