+++ .vscode/settings.json
@@ -1,0 +1,7 @@
+{
+  "workbench.colorCustomizations": {
+    "activityBar.background": "#5C0D22",
+    "titleBar.activeBackground": "#811230",
+    "titleBar.activeForeground": "#FEFBFC"
+  }
+}
