--- client/src/NewReviewForm.js
@@ -1,15 +1,0 @@
-import React from "react";
-
-const NewReviewForm = (props) => {
-
-
-
-
-    return (
-        <form>
-
-        </form>
-    )
-}
-
-export default NewReviewForm
