+++ client/src/services/isValid.js
@@ -1,0 +1,13 @@
+const isValid = (formPayload) => {
+  const requiredFields = ["rating", "vibe", "reviewText"];
+
+  for (let i = 0; i < requiredFields.length; i++) {
+    let currentFieldValue = requiredFields[i];
+    if (formPayload[currentFieldValue] === "") {
+      return false;
+    }
+  }
+  return true;
+}
+
+export default isValid;
